pyproject.toml
[project]
name = "texnostoremariupol"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask-bcrypt>=1.0.1",
    "flask-login>=0.6.3",
    "flask-migrate>=4.1.0",
    "flask-sqlalchemy>=3.1.1",
    "flask-wtf>=1.2.2",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.1",
    "wtforms>=3.2.1",
]
